widgets: Rework event handling a bit
This commit adds an event mask so that individual events can be masked
and unmasked.

It also clears up some of the technical debt by properly passing the
render configuration down the widget driver model. This fixes and
simplifies things a lot.

And because of these changes the pixmap widget is now much simpler that
it was before and easier to use.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>