widgets: Get rid of the negative notation for the redraw flags
The original idea I guess was that the data is initalized to zeroes so
the flags are set to zero from the start. But then the force redraw
flag was added and there is no point in having these this way and it's
way more confusing that I thought.

This also adds a gp_widget_should_redraw() helper that makes the code a
bit more readable.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>