grid: Fix GP_SELECT_IN handling
When the grid gets the GP_SELECT_IN for a first time the selected_col
and selected_row are both set to zero and if there is no selectable
widget (e.g. label) at that position in the grid we will wrongly return
that there is no widget to be selected, which may not be true.

So when we attempt to send a GP_SELECT_IN to the grid for a first time,
we have to traverse the grid with select_next() to see if there are any
selectable widgets widgets there.

This fixes a situation when there is a grid inside of a tabs widget with
a label at position 0,0 and textbox at 0,1. Previously the textbox
wasn't selectable until we clicked into the grid, which set the
selected_col and selected_row to grid coordinates of a selectable
widget.

This also means that we have to reset the selected flag when we remove
currently selected widget from the grid as well, so that we, on a next
GP_SELECT_IN, attempt to traverse again.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>